+++ ALZ.AKS/templates/terraform/modules/region/terraform.tf
@@ -17,6 +17,10 @@ terraform {
       source  = "Azure/azapi"
       version = "~> 2.0"
     }
+    random = {
+      source  = "hashicorp/random"
+      version = "~> 3.6"
+    }
     time = {
       source  = "hashicorp/time"
       version = "~> 0.12"